requirements.txt
Flask==2.0.3
Flask-RESTful==0.3.9
gunicorn==20.1.0
pyTelegramBotAPI==4.4.0
python-binance==1.0.15
python-dotenv==0.19.2
redis==4.1.4
regex==2022.1.18
requests==2.27.1
websockets==9.1
Werkzeug==2.0.3